--- .eclipse/configurations/HarderFarther - runGameTestServerSlim.launch
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<launchConfiguration type="org.eclipse.jdt.launching.localJavaApplication">
-  <stringAttribute key="org.eclipse.jdt.launching.PROJECT_ATTR" value="HarderFarther"/>
-  <stringAttribute key="org.eclipse.jdt.launching.MAIN_TYPE" value="cpw.mods.bootstraplauncher.BootstrapLauncher"/>
-  <stringAttribute key="org.eclipse.jdt.launching.VM_ARGUMENTS" value="-Dforge.logging.console.level=debug -Dforge.logging.markers=REGISTRIES -DlegacyClassPath.file=D:\Workspaces\Minecraft\Forge\1.20\HarderFarther\HarderFarther\build\classpath\runGameTestServer_minecraftClasspath.txt -DignoreList=bootstraplauncher,securejarhandler,asm-commons,asm-util,asm-analysis,asm-tree,asm,JarJarFileSystems,client-extra,fmlcore,javafmllanguage,lowcodelanguage,mclanguage,forge- -DmergeModules=jna-5.10.0.jar,jna-platform-5.10.0.jar -Dforge.enabledGameTestNamespaces=harderfarther -Djava.net.preferIPv6Addresses=system -p C:\Users\steph\.gradle\caches\modules-2\files-2.1\cpw.mods\bootstraplauncher\1.1.2\c546e00443d8432cda6baa1c860346980742628\bootstraplauncher-1.1.2.jar;C:\Users\steph\.gradle\caches\modules-2\files-2.1\cpw.mods\securejarhandler\2.1.6\66c15fc1f522b586476e9e4cccd0cbe192554e8a\securejarhandler-2.1.6.jar;C:\Users\steph\.gradle\caches\modules-2\files-2.1\org.ow2.asm\asm-commons\9.5\19ab5b5800a3910d30d3a3e64fdb00fd0cb42de0\asm-commons-9.5.jar;C:\Users\steph\.gradle\caches\modules-2\files-2.1\org.ow2.asm\asm-util\9.5\64b5a1fc8c1b15ed2efd6a063e976bc8d3dc5ffe\asm-util-9.5.jar;C:\Users\steph\.gradle\caches\modules-2\files-2.1\org.ow2.asm\asm-analysis\9.5\490bacc77de7cbc0be1a30bb3471072d705be4a4\asm-analysis-9.5.jar;C:\Users\steph\.gradle\caches\modules-2\files-2.1\org.ow2.asm\asm-tree\9.5\fd33c8b6373abaa675be407082fdfda35021254a\asm-tree-9.5.jar;C:\Users\steph\.gradle\caches\modules-2\files-2.1\org.ow2.asm\asm\9.5\dc6ea1875f4d64fbc85e1691c95b96a3d8569c90\asm-9.5.jar;C:\Users\steph\.gradle\caches\modules-2\files-2.1\net.minecraftforge\JarJarFileSystems\0.3.19\2464eb7d6b9ddb9db36a82cf8a95193e5c6fe020\JarJarFileSystems-0.3.19.jar --add-modules ALL-MODULE-PATH --add-opens java.base/java.util.jar=cpw.mods.securejarhandler --add-opens java.base/java.lang.invoke=cpw.mods.securejarhandler --add-exports java.base/sun.security.util=cpw.mods.securejarhandler --add-exports jdk.naming.dns/com.sun.jndi.dns=java.naming"/>
-  <stringAttribute key="org.eclipse.jdt.launching.PROGRAM_ARGUMENTS" value="--launchTarget forgegametestserveruserdev --gameDir . --fml.forgeVersion 46.0.10 --fml.mcVersion 1.20 --fml.forgeGroup net.minecraftforge --fml.mcpVersion 20230608.053357"/>
-  <stringAttribute key="org.eclipse.jdt.launching.WORKING_DIRECTORY" value="D:\Workspaces\Minecraft\Forge\1.20\HarderFarther\HarderFarther\run"/>
-  <mapAttribute key="org.eclipse.debug.core.environmentVariables">
-    <mapEntry key="MOD_CLASSES" value="harderfarther%%D:\Workspaces\Minecraft\Forge\1.20\HarderFarther\HarderFarther\bin\main;harderfarther%%D:\Workspaces\Minecraft\Forge\1.20\HarderFarther\HarderFarther\bin\main"/>
-    <mapEntry key="MCP_MAPPINGS" value="official_1.20"/>
-  </mapAttribute>
-</launchConfiguration>
