+++ gradle.properties
@@ -5,9 +5,9 @@ org.gradle.daemon=false
 
 
 ## Environment Properties
-minecraft_version=1.20
+minecraft_version=1.20.1
 minecraft_version_range=[1.20,1.21)
-forge_version=46.0.11
+forge_version=47.1.3
 forge_version_range=[46,)
 loader_version_range=[46,)
 mapping_channel=official
@@ -24,7 +24,7 @@ mod_name=Harder Farther
 # The license of the mod. Review your options at https://choosealicense.com/. All Rights Reserved is the default.
 mod_license=MIT
 # The mod version. See https://semver.org/
-mod_version=46.29.1
+mod_version=47.0.0
 # The group ID for the mod. It is only important when publishing as an artifact to a Maven repository.
 # This should match the base package used for the mod sources.
 # See https://maven.apache.org/guides/mini/guide-naming-conventions.html
@@ -33,3 +33,7 @@ mod_group_id=com.mactso.harderfarther
 mod_authors=MacTso, Lupicus
 # The description of the mod. This is a simple multiline text string that is used for display purposes in the mod list.
 mod_description=Makes minecraft harder and more rewarding based on distance from worldspawn
+
+
+## Extra dependencies
+terrablender_version=1.20.1-3.0.0.164
