+++ src/main/java/com/mactso/harderfarther/Main.java
@@ -3,6 +3,8 @@ package com.mactso.harderfarther;
 
 import java.lang.reflect.Field;
 
+import com.mactso.harderfarther.config.*;
+import com.mactso.harderfarther.events.*;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.jetbrains.annotations.NotNull;
@@ -11,17 +13,6 @@ import org.jetbrains.annotations.Nullable;
 import com.mactso.harderfarther.block.ModBlocks;
 import com.mactso.harderfarther.blockentities.ModBlockEntities;
 import com.mactso.harderfarther.command.HarderFartherCommands;
-import com.mactso.harderfarther.config.MyConfig;
-import com.mactso.harderfarther.events.BlockEvents;
-import com.mactso.harderfarther.events.ChunkEvent;
-import com.mactso.harderfarther.events.ExperienceDropEventHandler;
-import com.mactso.harderfarther.events.FogColorsEventHandler;
-import com.mactso.harderfarther.events.LivingEventMovementHandler;
-import com.mactso.harderfarther.events.MonsterDropEventHandler;
-import com.mactso.harderfarther.events.PlayerLoginEventHandler;
-import com.mactso.harderfarther.events.PlayerTickEventHandler;
-import com.mactso.harderfarther.events.PlayerTeleportHandler;
-import com.mactso.harderfarther.events.WorldTickHandler;
 import com.mactso.harderfarther.item.ModItems;
 import com.mactso.harderfarther.manager.GrimCitadelManager;
 import com.mactso.harderfarther.network.Register;
@@ -57,16 +48,21 @@ import net.minecraftforge.registries.RegisterEvent;
 public class Main {
 
 	    public static final String MODID = "harderfarther"; 
-		private static final Logger LOGGER = LogManager.getLogger();
+		public static final Logger LOGGER = LogManager.getLogger();
 		private static final int MAX_USABLE_VALUE = 16000000;  // you can subtract 1 from this number.
 
 	    
 	    public Main()
 	    {
 
-	    	Utility.debugMsg(0,MODID + ": Registering Mod.");
-	  		FMLJavaModLoadingContext.get().getModEventBus().register(this);
- 	        ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON,MyConfig.COMMON_SPEC );
+			Utility.debugMsg(0,MODID + ": Registering Mod.");
+			FMLJavaModLoadingContext.get().getModEventBus().register(this);
+			ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, PrimaryConfig.COMMON_SPEC );
+			BiomeConfig.initConfig();
+			OreConfig.initConfig();
+			StructureConfig.initConfig();
+			MobConfig.initConfig();
+			DimensionDifficultyOverridesConfig.initConfig();
 
 	    }
 
@@ -97,6 +93,8 @@ public class Main {
 				MinecraftForge.EVENT_BUS.register(new PlayerTeleportHandler());
 				MinecraftForge.EVENT_BUS.register(new LivingEventMovementHandler());
 				MinecraftForge.EVENT_BUS.register(new BlockEvents());
+				MinecraftForge.EVENT_BUS.register(new LivingEntitySpawnHandler());
+				DifficultyOverrideHandler.registerEvent();
 				fixAttributeMax();
  		}
