+++ src/main/java/com/mactso/harderfarther/config/DamageConfig.java
@@ -1,0 +1,7 @@
+package com.mactso.harderfarther.config;
+
+public class DamageConfig {
+
+    
+
+}
