+++ src/main/java/com/mactso/harderfarther/events/PlayerTickEventHandler.java
@@ -1,6 +1,6 @@
 package com.mactso.harderfarther.events;
 
-import com.mactso.harderfarther.config.MyConfig;
+import com.mactso.harderfarther.config.PrimaryConfig;
 import com.mactso.harderfarther.manager.HarderTimeManager;
 import net.minecraftforge.event.TickEvent.Phase;
 import net.minecraftforge.event.TickEvent.PlayerTickEvent;
@@ -13,7 +13,7 @@ public class PlayerTickEventHandler {
 
 		if ((event.phase != Phase.START))
 			return;
-		if (MyConfig.isMakeHarderOverTime()) {
+		if (PrimaryConfig.isMakeHarderOverTime()) {
 			HarderTimeManager.doScarySpookyThings(event.player);
 		}
