+++ src/main/java/com/mactso/harderfarther/manager/ChestLootManager.java
@@ -4,10 +4,11 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.StringTokenizer;
 
+import com.mactso.harderfarther.api.DifficultyCalculator;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import com.mactso.harderfarther.config.MyConfig;
+import com.mactso.harderfarther.config.PrimaryConfig;
 import com.mactso.harderfarther.utility.Utility;
 
 import net.minecraft.network.chat.Component;
@@ -75,7 +76,7 @@ public class ChestLootManager {
 	public static ItemStack doGetLootStack(ServerLevel level, Vec3 origin) {
 
 		RandomSource rand = level.getRandom();
-		float difficulty = HarderFartherManager.getDistanceDifficultyHere(level, origin);
+		float difficulty = DifficultyCalculator.getDistanceDifficultyHere(level, origin);
 
 		int rawroll = (int) (100 * difficulty);
 		int roll = rawroll;
@@ -89,7 +90,7 @@ public class ChestLootManager {
 		if (roll < 1)
 			roll = 1;
 		
-		int enchantmentLevel = 4 + MyConfig.getBonusLootEnchantmentLevelModifier();
+		int enchantmentLevel = 4 + PrimaryConfig.getBonusLootEnchantmentLevelModifier();
 
 		int chestLootRoll = Math.min(rand.nextInt(roll) + modifier, chestLootTable.size()-1);
 		ChestLootItem ci = chestLootTable.get(chestLootRoll);
