+++ src/main/java/com/mactso/harderfarther/manager/LootManager.java
@@ -9,7 +9,7 @@ import java.util.StringTokenizer;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import com.mactso.harderfarther.config.MyConfig;
+import com.mactso.harderfarther.config.PrimaryConfig;
 import com.mactso.harderfarther.item.ModItems;
 import com.mactso.harderfarther.utility.Utility;
 
@@ -180,7 +180,7 @@ public class LootManager {
 	
 	public static void doXPBottleDrop(Entity eventEntity, Collection<ItemEntity> eventItems, RandomSource rand) {
 		int d100 = (int) (Math.ceil(rand.nextDouble() * 100));
-		if (d100 < MyConfig.getOddsDropExperienceBottle()) {
+		if (d100 < PrimaryConfig.getOddsDropExperienceBottle()) {
 			Utility.debugMsg(1, "XP Bottle dropped with roll " + d100);
 			ItemStack itemStackToDrop;
 			itemStackToDrop = new ItemStack(Items.EXPERIENCE_BOTTLE, (int) 1);
