+++ src/main/java/com/mactso/harderfarther/mixinInterfaces/IExtendedServerWorld.java
@@ -1,0 +1,13 @@
+package com.mactso.harderfarther.mixinInterfaces;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public interface IExtendedServerWorld {
+
+    boolean areListInitialized();
+    void setListInitialized();
+    ArrayList<Float> getDifficultySectionNumbers();
+    ArrayList<List<String>> getDifficultySectionMobs();
+
+}
