+++ src/main/java/com/mactso/harderfarther/utility/Utility.java
@@ -4,7 +4,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.jetbrains.annotations.Nullable;
 
-import com.mactso.harderfarther.config.MyConfig;
+import com.mactso.harderfarther.config.PrimaryConfig;
 import com.mojang.brigadier.StringReader;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 
@@ -60,7 +60,7 @@ public class Utility {
 
 	public static void debugMsg(int level, String dMsg) {
 
-		if (MyConfig.getDebugLevel() > level - 1) {
+		if (PrimaryConfig.getDebugLevel() > level - 1) {
 			LOGGER.info("L" + level + ":" + dMsg);
 		}
 
@@ -68,7 +68,7 @@ public class Utility {
 
 	public static void debugMsg(int level, BlockPos pos, String dMsg) {
 
-		if (MyConfig.getDebugLevel() > level - 1) {
+		if (PrimaryConfig.getDebugLevel() > level - 1) {
 			LOGGER.info("L" + level + " (" + pos.getX() + "," + pos.getY() + "," + pos.getZ() + "): " + dMsg);
 		}
 
@@ -76,7 +76,7 @@ public class Utility {
 
 	public static void debugMsg(int level, LivingEntity le, String dMsg) {
 
-		if (MyConfig.getDebugLevel() > level - 1) {
+		if (PrimaryConfig.getDebugLevel() > level - 1) {
 			LOGGER.info("L" + level + " (" 
 					+ le.blockPosition().getX() + "," 
 					+ le.blockPosition().getY() + ","
