+++ src/main/resources/META-INF/accesstransformer.cfg
@@ -1,0 +1,6 @@
+public net.minecraft.world.level.biome.Climate$RTree
+public net.minecraft.world.level.biome.Climate$RTree$Node
+public net.minecraft.world.level.biome.Climate$RTree$Node *()
+public net.minecraft.world.level.biome.Climate$RTree$Leaf *()
+public net.minecraft.world.level.biome.Climate$RTree$SubTree *()
+public net.minecraft.world.level.biome.Climate$DistanceMetric
